Add max-width formatter for string keys
This gives flexibility that might be desirable for "format_current",
e.g. to limit the artist and album to at most 20 characters each:
format_current= %.20a / %.20l / %n - %t%= %y

Other types are out of scope, and the intent is not to have full printf
compatibility. We are already incompatible in various ways.

Closes #1004.